icer: deep-sample 9..=16-bit 2-D grayscale — the §II.C 12-bit MER operating point end to end
IcerPixelFormat::GrayDeep { bits } rides the plane container with a
new format tag 2 + one bit-depth byte (the 12-byte segment header has
no free field, and a bare deep stream would be indistinguishable from
an 8-bit one). Inside is a normal single-plane segment stream whose
coefficients span the deeper range: the level shift generalises to
2^(n-1), the decode clamp to [0, 2^n - 1], placeholder fills to the
n-bit midpoint, and the §III.D raw path ships little-endian sample
pairs. §II.C Table 4 guarantees 16-bit input can never overflow the
i32 coefficient words.

Every existing wire form is unchanged: Gray8 streams stay bare
byte-for-byte, colour containers keep tag 1. Deep composes with all
seven §II.A filters (lossless pinned at every depth 9..=16), row-strip
and §V.B transform-domain segments, both entropy backends, §III.A
priority interleaving, §VI.A min-loss, §VI.B quotas (the 9 container
framing bytes are charged against the budget so the hard cap bounds
the total output), ROI priorities, the §III.D fallback, lenient decode
(deep midpoint fill), DecodeLimits, and quality-target rate-control.
PSNR/SSIM peaks follow the §VII definition (2^b - 1); ImageStats scans
deep samples down-shifted to 8 bits so the filter-selection thresholds
keep their calibration. 20 new integration tests + container unit
pins.

+++ tests/distortion_report.rs
@@ -77,12 +77,12 @@ fn distortion_report_constant_offset_metrics_are_exact() {
     let h = original.height as usize;
     let mut sq = 0.0f64;
     let mut ab = 0.0f64;
-    let mut mx = 0u8;
+    let mut mx = 0u16;
     for y in 0..h {
         for x in 0..w {
             let o = original.planes[0].data[y * stride_o + x] as i32;
             let d = decoded.planes[0].data[y * stride_d + x] as i32;
-            let a = (o - d).unsigned_abs() as u8;
+            let a = (o - d).unsigned_abs() as u16;
             if a > mx {
                 mx = a;
             }
